backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
SQLAlchemy==2.0.20
asyncpg==0.31.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
pydantic==2.12.4
python-dotenv==1.0.0
email-validator==2.3.0
httpx==0.28.1
cryptography==46.0.1
